Move files into main project
