+++ Lombiq.Tests.UI/Services/OrchardCoreUITestExecutorConfiguration.cs
@@ -1,4 +1,5 @@
 using Lombiq.Tests.UI.Extensions;
+using Lombiq.Tests.UI.SqlQueryMonitoring.Services;
 using Lombiq.Tests.UI.Helpers;
 using Lombiq.Tests.UI.SecurityScanning;
 using Lombiq.Tests.UI.Services.GitHub;
