+++ Lombiq.Tests.UI/SqlQueryMonitoring/Exceptions/SqlQueryMonitoringAssertionException.cs
@@ -1,8 +1,9 @@
-using Lombiq.Tests.UI.AppExtensions.SqlQueryMonitoring;
-using Lombiq.Tests.UI.Services;
+using Lombiq.Tests.UI.Exceptions;
+using Lombiq.Tests.UI.SqlQueryMonitoring;
+using Lombiq.Tests.UI.SqlQueryMonitoring.Services;
 using System;
 
-namespace Lombiq.Tests.UI.Exceptions;
+namespace Lombiq.Tests.UI.SqlQueryMonitoring.Exceptions;
 
 public class SqlQueryMonitoringAssertionException : Exception, IAssertionException
 {
