+++ Lombiq.Tests.UI/SqlQueryMonitoring/SqlQueryMonitoringContext.cs
@@ -1,14 +1,14 @@
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 
-namespace Lombiq.Tests.UI.AppExtensions.SqlQueryMonitoring;
+namespace Lombiq.Tests.UI.SqlQueryMonitoring;
 
 public sealed class SqlQueryMonitoringContext : ISqlQueryMonitoringContext
 {
     private readonly ConcurrentQueue<SqlQueryExecutionEntry> _executions = new();
 
     public IReadOnlyList<SqlQueryExecutionEntry> Executions
-        => _executions.ToArray();
+        => [.. _executions];
 
     public void RecordExecution(SqlQueryExecutionEntry entry)
     {
